- Updated file structure.- Fixed bug in registration.- Updated post and user providers.
